requirements_base.txt
boto==2.42.0
django-localflavor==1.3
Django==1.8.8
otree-core[mturk]==1.4.17
